replica: allow fractional listing economics
- make trade product display amounts decimal-friendly
- preserve fractional listing exact values during ingest
- add fractional replica ingest coverage
- keep exact economics strings as signed order source

+++ crates/replica_db_schema/src/models/trade_product.rs
@@ -18,7 +18,7 @@ pub struct TradeProduct {
     pub lot: String,
     pub profile: String,
     pub year: i64,
-    pub qty_amt: i64,
+    pub qty_amt: f64,
     pub qty_amt_exact: Option<String>,
     pub qty_unit: String,
     pub qty_label: Option<String>,
@@ -26,7 +26,7 @@ pub struct TradeProduct {
     pub price_amt: f64,
     pub price_amt_exact: Option<String>,
     pub price_currency: String,
-    pub price_qty_amt: u32,
+    pub price_qty_amt: f64,
     pub price_qty_amt_exact: Option<String>,
     pub price_qty_unit: String,
     pub listing_addr: Option<String>,
@@ -45,7 +45,7 @@ pub struct ITradeProductFields {
     pub lot: String,
     pub profile: String,
     pub year: i64,
-    pub qty_amt: i64,
+    pub qty_amt: f64,
     pub qty_amt_exact: String,
     pub qty_unit: String,
     #[cfg_attr(feature = "ts-rs", ts(optional, type = "string | null"))]
@@ -55,7 +55,7 @@ pub struct ITradeProductFields {
     pub price_amt: f64,
     pub price_amt_exact: String,
     pub price_currency: String,
-    pub price_qty_amt: u32,
+    pub price_qty_amt: f64,
     pub price_qty_amt_exact: String,
     pub price_qty_unit: String,
     #[cfg_attr(feature = "ts-rs", ts(optional, type = "string | null"))]
@@ -141,7 +141,7 @@ pub struct ITradeProductFieldsFilter {
     #[cfg_attr(feature = "ts-rs", ts(optional))]
     pub year: Option<i64>,
     #[cfg_attr(feature = "ts-rs", ts(optional))]
-    pub qty_amt: Option<i64>,
+    pub qty_amt: Option<f64>,
     #[cfg_attr(feature = "ts-rs", ts(optional))]
     pub qty_amt_exact: Option<String>,
     #[cfg_attr(feature = "ts-rs", ts(optional))]
@@ -157,7 +157,7 @@ pub struct ITradeProductFieldsFilter {
     #[cfg_attr(feature = "ts-rs", ts(optional))]
     pub price_currency: Option<String>,
     #[cfg_attr(feature = "ts-rs", ts(optional))]
-    pub price_qty_amt: Option<u32>,
+    pub price_qty_amt: Option<f64>,
     #[cfg_attr(feature = "ts-rs", ts(optional))]
     pub price_qty_amt_exact: Option<String>,
     #[cfg_attr(feature = "ts-rs", ts(optional))]
